add ring plot
A ring plot is a polar plot that has a minimal radius.
It uses polar coordinates and transforms the min and max radial values
to the inner and outer rings respectively.
Data is hold in line X and Y not C as for polar data.
Radial data can be negative.